renaming file names to match component names
