+++ protocol-designer/src/components/StepEditForm/fields/ChangeTipField/ChangeTip.js
