+++ protocol-designer/src/components/StepEditForm/fields/PathField/index.js
