+++ protocol-designer/src/components/StepEditForm/fields/TipPositionField/TipPositionInput.css
