+++ protocol-designer/src/components/StepEditForm/fields/TipPositionField/TipPositionModal.js
