+++ protocol-designer/src/components/StepEditForm/fields/TipPositionField/TipPositionZAxisViz.js
