+++ protocol-designer/src/components/StepEditForm/fields/TipPositionField/index.js
