+++ protocol-designer/src/components/StepEditForm/fields/TipPositionField/utils.js
