+++ protocol-designer/src/components/StepEditForm/fields/VolumeField.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { FormGroup, HoverTooltip } from '@opentrons/components'
 import i18n from '../../../localization'
 import { getTooltipForField } from '../utils'
-import { TextField } from './Text'
+import { TextField } from './TextField'
 import type { StepType } from '../../../form-types'
 import type { FocusHandlers } from '../types'
 import styles from '../StepEditForm.css'
