+++ protocol-designer/src/components/StepEditForm/fields/index.js
@@ -2,22 +2,22 @@
 
 /* Generic Fields */
 
-export { CheckboxRowField } from './CheckboxRow'
-export { RadioGroupField } from './RadioGroup'
-export { TextField } from './Text'
+export { CheckboxRowField } from './CheckboxRowField'
+export { RadioGroupField } from './RadioGroupField'
+export { TextField } from './TextField'
 
 /* Specialized Fields */
 
 export { ConditionalOnField } from './ConditionalOnField'
-export { WellSelectionField } from './WellSelection'
-export { WellOrderField } from './WellOrder'
-export { TipPositionField } from './TipPosition'
-export { FlowRateField } from './FlowRate'
-export { BlowoutLocationField } from './BlowoutLocation'
-export { ChangeTipField } from './ChangeTip'
-export { DisposalVolumeField } from './DisposalVolume'
-export { LabwareField } from './Labware'
-export { PathField } from './Path'
-export { PipetteField } from './Pipette'
-export { VolumeField } from './Volume'
-export { StepFormDropdown } from './StepFormDropdown'
+export { WellSelectionField } from './WellSelectionField'
+export { WellOrderField } from './WellOrderField'
+export { TipPositionField } from './TipPositionField'
+export { FlowRateField } from './FlowRateField'
+export { BlowoutLocationField } from './BlowoutLocationField'
+export { ChangeTipField } from './ChangeTipField'
+export { DisposalVolumeField } from './DisposalVolumeField'
+export { LabwareField } from './LabwareField'
+export { PathField } from './PathField'
+export { PipetteField } from './PipetteField'
+export { VolumeField } from './VolumeField'
+export { StepFormDropdown } from './StepFormDropdownField'
